mobile layout, music window, + click zorder
created a basic and readable mobile layout.

added a little window with my favorite artists [though i wanted this to be a volt.fm widget, it doesn't seem they have an api, or at least a public one]. [this is at least a good experiment in what windows i can make.]

and now when you click any window anywhere, it moves it to the top of the zorder.

also:

removed fontawesome [not needed? pixels are cooler.]

added link to git repo and issue reports.

edited link/nav styles.

added navlink hrefs.

began readme.